+++ crates/parol/CHANGELOG.md
@@ -42,6 +42,8 @@ which closes issue [#21](https://github.com/jsinger67/parol/issues/21)
   * The parser detects conflicting token aliases pairwise and issues a dedicated error message
 * Extended enhancement from issue [#19](https://github.com/jsinger67/parol/issues/19) to
 `line_comment` and `block_comment` directives
+* Fixed [#39](https://github.com/jsinger67/parol/issues/39)
+  * Empty scanner states are detected at the end of the parse process
 
 ## v0.14.0 - 2022-11-18